% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain_mcce}
\alias{explain_mcce}
\title{Fit autoregressive model to the training data}
\usage{
explain_mcce(
  model,
  x_explain,
  x_train,
  predict_model = shapr::predict_model,
  fixed_features,
  c_int = c(0.5, 1),
  fit.autoregressive_model = "ctree",
  fit.decision = TRUE,
  fit.seed = NULL,
  generate.K = 1000,
  generate.seed = NULL,
  process.measures = c("validation", "L0", "L1"),
  process.return_best_k = 1,
  process.remove_invalid = TRUE,
  process.sort_by_measures_order = TRUE
)
}
\arguments{
\item{model}{The model whose predictions we want to explain.}

\item{x_explain}{Matrix or data.frame/data.table.
Contains the data that we want to generate counterfactuals for.}

\item{x_train}{Matrix or data.frame/data.table.
Contains the data used to train the autoregressive feature dependence model.}

\item{predict_model}{Function.
The prediction function.
The function must have two arguments, \code{model} and \code{newdata} which specify, respectively, the model
and a data.frame/data.table to compute predictions for. The function must give the prediction as a numeric vector.}

\item{fixed_features}{Character vector.
Names of features to fix in counterfactual generation.}

\item{c_int}{Numeric vector (length 2).
Specifies the range of prediction values corresponding to the desired decision}

\item{fit.autoregressive_model}{Character.
Specifies the name of the autoregressive model used to fit the data.}

\item{fit.decision}{Logical.
Whether to include the decision threshold as a binary features to improve the efficiency of MCCE.}

\item{fit.seed}{Positive integer.
Specifies the seed used when fitting the autoregressive feature dependence model.
If \code{NULL} the seed will be inherited from the calling environment.}

\item{generate.K}{Numeric.
Number of potential counterfactuals to generate per row in x_explain.}

\item{generate.seed}{Positive integer.
Specifies the seed used when generating the potential counterfactuals.
If \code{NULL} the seed will be inherited from the calling environment.}

\item{process.measures}{Character vector.
Indicates the measures (in the given order) that are applied to the simulate data.}

\item{process.return_best_k}{Integer.
How many counterfactuals should be generated per prediction to explain (at max).}

\item{process.remove_invalid}{Logical.
Indicates whether invalid counterfactuals should be removed from the returned counterfactual list.}

\item{process.sort_by_measures_order}{Logical.
Indicates whether the counterfactuals should be sorted.}
}
\value{
List. The counterfactuals for the predictions we explain.
}
\description{
Fit autoregressive model to the training data
}
