% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate potential counterfactuals from fitted autoregressive model}
\usage{
generate(x_explain, fit_object, K = 1000, seed = NULL)
}
\arguments{
\item{x_explain}{Matrix or data.frame/data.table.
Contains the data that we want to generate counterfactuals for.}

\item{fit_object}{List.
The output from calling the fit() function.}

\item{K}{Numeric.
Number of potential counterfactuals to generate per row in x_explain.}

\item{seed}{Positive integer.
Specifies the seed used when generating the potential counterfactuals.
If \code{NULL} the seed will be inherited from the calling environment.}
}
\value{
List. The generated potential counterfactuals and the timing of this function.
}
\description{
Generate potential counterfactuals from fitted autoregressive model
}
