% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcceR-package.R
\docType{package}
\name{mcceR-package}
\alias{mcceR}
\alias{mcceR-package}
\title{mcceR: Monte Carlo sampling of realistic Counterfactual Explanations for tabular data}
\description{
mcceR generate counterfactual explanations for any tabular data model using the method described in Redelmeier, A., Jullum, M., Aas, K. and Løland, A. (2021) \href{https://arxiv.org/abs/2111.09790}{arXiv:2111.09790}. A Python wrapper (mcceRpy) is also available to explain models fitted in Python.
}
\author{
\strong{Maintainer}: Martin Jullum \email{Martin.Jullum@nr.no} (\href{https://orcid.org/0000-0003-3908-5155}{ORCID})

Authors:
\itemize{
  \item Annabelle Redelmeier \email{Annabelle.Redelmeier@nr.no}
}

Other contributors:
\itemize{
  \item Anders Løland \email{Anders.Loland@nr.no} [contributor]
  \item Norsk Regnesentral [copyright holder, funder]
}

}
\keyword{internal}
