% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict_model}
\alias{predict_model}
\alias{predict_model.default}
\alias{predict_model.xgb.Booster}
\alias{predict_model.ranger}
\alias{predict_model.lm}
\alias{predict_model.gam}
\alias{predict_model.glm}
\title{Generate predictions for input data with specified model}
\usage{
predict_model(x, newdata)

\method{predict_model}{default}(x, newdata)

\method{predict_model}{xgb.Booster}(x, newdata)

\method{predict_model}{ranger}(x, newdata)

\method{predict_model}{lm}(x, newdata)

\method{predict_model}{gam}(x, newdata)

\method{predict_model}{glm}(x, newdata)
}
\arguments{
\item{x}{Model object for the model to be explained.}

\item{newdata}{A data.frame/data.table with the features to predict from.}
}
\value{
Numeric. Vector of size equal to the number of rows in \code{newdata}.
}
\description{
Performs prediction of response
\code{\link[stats:lm]{stats::lm()}},
\code{\link[stats:glm]{stats::glm()}},
\code{\link[ranger:ranger]{ranger::ranger()}},
\code{\link[mgcv:gam]{mgcv::gam()}} and
\code{\link[xgboost:xgb.train]{xgboost::xgb.train()}} with binary or continuous
response. See details for more information.

NOTE: You should never need to call this function explicitly.
It is exported just to be easier accessible for users, see details.
}
\details{
The following models are currently supported:
\itemize{
\item \code{\link[stats:lm]{stats::lm()}}
\item \code{\link[stats:glm]{stats::glm()}}
\item \code{\link[ranger:ranger]{ranger::ranger()}}
\item \code{\link[mgcv:gam]{mgcv::gam()}}
\item \code{\link[xgboost:xgb.train]{xgboost::xgb.train()}}
}

If you have a binary classification model we'll always return the probability prediction
for a single class.

If you are explaining a model not supported natively, you need to create the \verb{[predict_model()]} function yourself,
and pass it on to as an argument to \verb{[explain()]}.
}
\keyword{internal}
