% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Generate potential counterfactuals from fitted autoregressive model}
\usage{
process(
  x_sim,
  pred_sim,
  x_explain,
  fit_object,
  measures = c("validation", "L0", "L1"),
  remove_invalid = TRUE,
  return_best_k = 1,
  sort_by_measures_order = TRUE
)
}
\arguments{
\item{x_sim}{data.table.
Contains the generated potential  counterfactuals, including an id-column (id_explain).
The output from generate()$x_sim}

\item{pred_sim}{Numeric vector.
The prediction score for the data in x_sim.}

\item{x_explain}{Matrix or data.frame/data.table.
Contains the data that we want to generate counterfactuals for.}

\item{fit_object}{List.
The output from calling the fit() function.}

\item{measures}{Character vector.
Indicates the measures (in the given order) that are applied to the simulate data.}

\item{remove_invalid}{Logical.
Indicates whether invalid counterfactuals should be removed from the returned counterfactual list.}

\item{return_best_k}{Integer.
How many counterfactuals should be generated per prediction to explain (at max).}

\item{sort_by_measures_order}{Logical.
Indicates whether the counterfactuals should be sorted.}
}
\value{
List.
Procesed counterfactuals, and timing
}
\description{
Generate potential counterfactuals from fitted autoregressive model
}
