% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit}
\alias{fit}
\title{Fit and sample a set of ctrees based on mutable (not fixed) features}
\usage{
fit(
  x_train,
  pred_train,
  fixed_features,
  c_val = mean(pred_train),
  autoregressive_model = "ctree",
  seed = NULL
)
}
\arguments{
\item{x_train}{Data.frame or data.table of features and response}

\item{pred_train}{Numeric vector. Specifies the prediction value for the observations in x_train.}

\item{fixed_features}{Character vector. Names of features to fix in counterfactual generation.}

\item{c_val}{Numeric. Specifying the decision threshold to exceed to achieve the right decison}

\item{autoregressive_model}{Character. Specifies the name of the autoregressive model used to fit the data.}

\item{seed}{Numeric.}
}
\value{
List. Includes \code{synData} of class "synds" which includes the counterfactual exaplanations.
}
\description{
Fit and sample a set of ctrees based on mutable (not fixed) features
}
