% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit autoregressive model to the training data}
\usage{
fit(
  x_train,
  pred_train,
  fixed_features,
  c_int = c(mean(pred_train), 1),
  featuremodel_object = NULL,
  autoregressive_model = "ctree",
  seed = NULL,
  decision = TRUE,
  ...
)
}
\arguments{
\item{x_train}{Matrix or data.frame/data.table.
Contains the data used to train the autoregressive feature dependence model.}

\item{pred_train}{Numeric vector.
The prediction score for the data in x_train.}

\item{fixed_features}{Character vector.
Names of features to fix in counterfactual generation.}

\item{c_int}{Numeric vector (length 2).
Specifies the range of prediction values corresponding to the desired decision}

\item{autoregressive_model}{Character.
Specifies the name of the autoregressive model used to fit the data.}

\item{seed}{Positive integer.
Specifies the seed used when fitting the autoregressive feature dependence model.
If \code{NULL} the seed will be inherited from the calling environment.}

\item{decision}{Logical.
Whether to include the decision threshold as a binary features to improve the efficiency of MCCE.}

\item{...}{Additional arguments passed to \code{\link[party:ctree]{party::ctree()}} or
\code{\link[rpart:rpart]{rpart::rpart()}}}
}
\value{
List. Includes featuremodel_list with the fitted models
}
\description{
Fit autoregressive model to the training data
}
